+++ lib/.gitignore
@@ -58,7 +58,7 @@ javaManager.js
 jarManager.js
 javaTypes.js
 tomcatManager.js
-sync.js
+sync*.js
 
 # This program and the accompanying materials are
 # made available under the terms of the Eclipse Public License v2.0 which accompanies
