+++ lib/webapp.js
@@ -36,6 +36,8 @@ const expressStaticGzip = require("express-static-gzip");
 const os = require('os');
 const semver = require('semver');
 const ipaddr = require('ipaddr.js');
+const SyncService = require('./sync-service').SyncService;
+const SyncClient = require('./sync-client').SyncClient;
 const sync = require('./sync');
 
 /**
@@ -681,7 +683,7 @@ const staticHandlers = {
     const router = express.Router();
     webApp.expressWs.applyTo(router);
     router.ws('/sync', (ws, req) => {
-      new sync.SyncEndpoint(ws, req);
+      new SyncService(ws, req);
     });
     return router;
   },
@@ -1802,9 +1804,11 @@ WebApp.prototype = {
     console.log(`installSyncClient port ${port}`);
     if (port !== 8544) {
       console.log('installSyncClient backup');
-      const client = sync.SyncClient();
+      sync.setAsBackup();
+      const client = new SyncClient();
     } else {
       console.log('installSyncClient master');
+      sync.setAsMaster();
     }
   },
